@asomov 2024-08-16
the change in SafeConstructor is indeed an improvement.
I am not sure the changes in Scanner provide a significant improvement - for sure it makes the code less readable
(object creation is a very cheap operation in JVM)